Extract Dashboard subcomponents
