+++ frontend/src/components/dashboard/index.ts
@@ -1,3 +1,17 @@
+export { DashboardDemoBanner, DashboardSetupEmptyState } from "./DashboardEmptyState"
+export { DashboardHero } from "./DashboardHero"
+export { DashboardMetricGrid } from "./DashboardMetricGrid"
+export { DashboardRemediationSection } from "./DashboardRemediationSection"
+export { DashboardSidePanel } from "./DashboardSidePanel"
+export { DashboardSignalOverview } from "./DashboardSignalOverview"
 export { ProviderFreshnessPanel } from "./ProviderFreshnessPanel"
 export { TopRemediationQueue } from "./TopRemediationQueue"
 export { RiskOperationsDashboard } from "./RiskOperationsDashboard"
+export type {
+  DashboardMetricSummary,
+  DashboardRunFact,
+  DashboardRunRange,
+  DashboardSignalCounts,
+  QueueFilterState,
+  RiskOperationsDashboardProps,
+} from "./dashboard-model"
